refactor: move mock-only tests out of common directory
Move tests with mock-specific dependencies to e2e/mock/degraded-mode/
to comply with ESLint rule banning test.skip() calls:

1. Created authentication-query-params.spec.ts for query param redirect test
2. Created diff-view-mock-only.spec.ts for:
   - Syntax highlighting preserved (Issue #131)
   - Virtualized diff navigation (500+ lines)

Removed test.skip() calls from common directory - tests are now properly
organized by their execution requirements.

Test counts:
- Mock: 46 passed
- Prod: 34 passed
